requirements.txt
fastapi==0.88.0
pydantic==1.10.4
iso8601==1.1.0
uvicorn~=0.20.0
ydb==2.12.0
requests==2.28.2
six
python-dotenv
loguru